packages/imaging-common/pyproject.toml
[project]
name = "imaging-common"
version = "0.1.0"
description = "Shared infrastructure for imaging pipeline services"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary",
    "pyyaml",
    "requests",
    "croniter",
    "pydantic-settings",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/imaging_common"]
